Fix some signed/unsigned warnings
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20824 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/tests/asn.1/trval.c
@@ -145,7 +145,7 @@ int trval(fin, fout)
 	p = (unsigned char *)malloc(maxlen);
 	len = 0;
 	while ((cc = fgetc(fin)) != EOF) {
-		if (len == maxlen) {
+		if ((unsigned int) len == maxlen) {
 			maxlen += BUFSIZ;
 			p = (unsigned char *)realloc(p, maxlen);
 		}
